pagesからappにファイルを移動
