--- pages/index.tsx
@@ -1,43 +1,0 @@
-import Container from "../components/container";
-import MoreStories from "../components/more-stories";
-import Layout from "../components/layout";
-import { getAllPosts } from "../lib/api";
-import Head from "next/head";
-import Post from "../interfaces/post";
-
-type Props = {
-  allPosts: Post[];
-};
-
-export default function Index({ allPosts }: Props) {
-  return (
-    <>
-      <Layout>
-        <Head>
-          <title>{"yummy yummy bread"}</title>
-        </Head>
-        <Container>
-          {allPosts.length > 0 && (
-            <MoreStories posts={allPosts} title="Articles" />
-          )}
-        </Container>
-      </Layout>
-    </>
-  );
-}
-
-export const getStaticProps = async () => {
-  const allPosts = getAllPosts([
-    "title",
-    "date",
-    "slug",
-    "author",
-    "coverImage",
-    "excerpt",
-    "tags",
-  ]);
-
-  return {
-    props: { allPosts },
-  };
-};
